+++ test/sync/src/androidTest/kotlin/io/realm/test/mongodb/shared/SyncConfigTests.kt
@@ -35,6 +35,7 @@ import io.realm.test.util.TestHelper.getRandomKey
 import io.realm.test.util.TestHelper.randomEmail
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
+import kotlin.test.Ignore
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -235,29 +236,59 @@ class SyncConfigTests {
             user = user,
             partitionValue = partitionValue
         ).build()
-        assertEquals(config.partitionValue.asString(), partitionValue)
+        assertEquals(config.user, user)
+        assertTrue(config.schema.containsAll(setOf(ParentPk::class, ChildPk::class)))
+        // assertEquals(config.partitionValue.asString(), partitionValue)
     }
 
     @Test
-    fun name_notSpecifiedWithStringPartitionValue() {
-        val fileName = "s_mypartition.realm"
+    fun name_partitionBasedDefaultName() {
         val user: User = createTestUser()
-        val config: SyncConfiguration = SyncConfiguration.Builder(user, "mypartition", setOf())
-            .build()
-        val suffix = "/mongodb-realm/${user.app.configuration.appId}/${user.identity}/$fileName"
+        // Long
+        verifyName(
+            SyncConfiguration.Builder(user, 1234L, setOf()),
+            "l_1234.realm"
+        )
+
+        // Int
+        verifyName(
+            SyncConfiguration.Builder(user, 123 as Int, setOf()),
+            "i_123.realm"
+        )
+
+        // String
+        verifyName(
+            SyncConfiguration.Builder(user, "mypartition", setOf()),
+            "s_mypartition.realm"
+        )
+    }
+
+    private fun verifyName(builder: SyncConfiguration.Builder, expectedFileName: String) {
+        val config = builder.build()
+        val suffix = "/mongodb-realm/${config.user.app.configuration.appId}/${config.user.identity}/$expectedFileName"
         assertTrue(config.path.contains(suffix), "${config.path} failed.")
-        assertEquals(fileName, config.name)
+        assertEquals(expectedFileName, config.name)
     }
 
     @Test
-    fun name_notSpecifiedWithLongPartitionValue() {
-        val fileName = "l_1234567890.realm"
+    fun nullPartitionValue() {
         val user: User = createTestUser()
-        val config: SyncConfiguration = SyncConfiguration.Builder(user, 1234567890L, setOf())
-            .build()
-        val suffix = "/mongodb-realm/${user.app.configuration.appId}/${user.identity}/$fileName"
-        assertTrue(config.path.contains(suffix), "${config.path} failed.")
-        assertEquals(fileName, config.name)
+        val configs = listOf<SyncConfiguration>(
+            SyncConfiguration.with(user, null as String?, setOf()),
+            SyncConfiguration.with(user, null as Int?, setOf()),
+            SyncConfiguration.with(user, null as Long?, setOf()),
+            // SyncConfiguration.with(user, null as ObjectId?),
+            // SyncConfiguration.with(user, null as UUID?),
+            SyncConfiguration.Builder(user, null as String?, setOf()).build(),
+            SyncConfiguration.Builder(user, null as Int?, setOf()).build(),
+            SyncConfiguration.Builder(user, null as Long?, setOf()).build(),
+            // SyncConfiguration.Builder(user, null as ObjectId?).build()
+            // SyncConfiguration.Builder(user, null as UUID?).build()
+        )
+
+        configs.forEach { config ->
+            assertTrue(config.path.endsWith("/null.realm"), "${config.path} failed")
+        }
     }
 
     @Test
@@ -352,13 +383,13 @@ class SyncConfigTests {
 //        assertNotNull(config.rxFactory)
 //    }
 //
-//    @Test
-//    fun toString_nonEmpty() {
-//        val user: User = createTestUser(app)
-//        val config: SyncConfiguration = SyncConfiguration.defaultConfig(user, DEFAULT_PARTITION)
-//        val configStr = config.toString()
-//        assertTrue(configStr.isNotEmpty())
-//    }
+    @Test
+    fun toString_nonEmpty() {
+        val user: User = createTestUser()
+        val config: SyncConfiguration = SyncConfiguration.with(user, partitionValue, setOf())
+        val configStr = config.toString()
+        assertTrue(configStr.isNotEmpty())
+    }
 //
 //    // Check that it is possible for multiple users to reference the same Realm URL while each user still use their
 //    // own copy on the filesystem. This is e.g. what happens if a Realm is shared using a PermissionOffer.
@@ -382,13 +413,15 @@ class SyncConfigTests {
 //        assertNotEquals(config1.path, config2.path)
 //    }
 //
-//    @Test
-//    fun defaultConfiguration_throwsIfNotLoggedIn() {
-//        // TODO Maybe we could avoid registering a real user
-//        val user = app.registerUserAndLogin(TestHelper.getRandomEmail(), "123456")
-//        user.logOut()
-//        assertFailsWith<IllegalArgumentException> { SyncConfiguration.defaultConfig(user, DEFAULT_PARTITION) }
-//    }
+    @Test
+    fun with_throwsIfNotLoggedIn() = runBlocking {
+        val user: User = createTestUser()
+        user.logOut()
+        assertFailsWith<IllegalArgumentException> { SyncConfiguration.with(user, "string", setOf()) }
+        assertFailsWith<IllegalArgumentException> { SyncConfiguration.with(user, 123 as Int, setOf()) }
+        assertFailsWith<IllegalArgumentException> { SyncConfiguration.with(user, 123L, setOf()) }
+        Unit
+    }
 //
 //    @Test
 //    @Ignore("Not implemented yet")
@@ -410,6 +443,7 @@ class SyncConfigTests {
 //    fun getUrlPrefix() {
 //    }
 
+    @Ignore // Wait for https://github.com/realm/realm-kotlin/issues/648
     @Test
     fun getPartitionValue() {
         val user = createTestUser()
@@ -418,7 +452,8 @@ class SyncConfigTests {
             user = user,
             partitionValue = partitionValue
         ).build()
-        assertEquals(partitionValue, config.partitionValue.asString())
+        // Disabled until we have a proper BSON API
+        // assertEquals(DEFAULT_PARTITION_VALUE, config.partitionValue.asString())
     }
 
 //    @Test
@@ -469,25 +504,6 @@ class SyncConfigTests {
 //        }
 //    }
 //
-//    @Test
-//    fun nullPartitionValue() {
-//        val user: User = createTestUser(app)
-//
-//        val configs = listOf<SyncConfiguration>(
-//            SyncConfiguration.defaultConfig(user, null as String?),
-//            SyncConfiguration.defaultConfig(user, null as Int?),
-//            SyncConfiguration.defaultConfig(user, null as Long?),
-//            SyncConfiguration.defaultConfig(user, null as ObjectId?),
-//            SyncConfiguration.Builder(user, null as String?).build(),
-//            SyncConfiguration.Builder(user, null as Int?).build(),
-//            SyncConfiguration.Builder(user, null as Long?).build(),
-//            SyncConfiguration.Builder(user, null as ObjectId?).build()
-//        )
-//
-//        configs.forEach { config ->
-//            assertTrue(config.path.endsWith("/null.realm"))
-//        }
-//    }
 //
 //    @Test
 //    fun loggedOutUsersThrows() {